+++ miden-vm/benches/program_execution_fast.rs
@@ -56,10 +56,9 @@ fn program_execution_fast(c: &mut Criterion) {
                                 .with_library(&CoreLibrary::default())
                                 .unwrap();
 
-                            let processor = FastProcessor::new_with_advice_inputs(
-                                stack_inputs,
-                                advice_inputs.clone(),
-                            );
+                            let processor = FastProcessor::builder(stack_inputs)
+                                .with_advice(advice_inputs.clone())
+                                .build();
 
                             (host, program.clone(), processor)
                         },