+++ processor/src/processor/operations/fri_ops/tests.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use miden_core::{
     Felt,
     field::{BasedVectorSpace, Field, PrimeCharacteristicRing, QuadFelt, TwoAdicField},