+++ processor/src/processor/operations/sys_ops/tests.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use alloc::vec;
 
 use miden_core::{