+++ processor/src/processor/operations/u32_ops/tests.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use alloc::vec::Vec;
 
 use miden_core::{